# Conflicts:
#	reports/libopus_type_math_parity_audit.md

+++ celt/hybrid_encode_helpers.go
@@ -603,7 +603,7 @@ func (e *Encoder) SetCoarseEnergyAvailableBytes(bytes int) {
 	if bytes < 0 {
 		bytes = 0
 	}
-	e.coarseAvailableBytes = bytes
+	e.coarseAvailableBytes = int32(bytes)
 }
 
 // MDCTScratch computes the MDCT using the encoder's pre-allocated scratch buffers.
